+++ rotors_description/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package rotors_description
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.0.10 (2020-04-14)
+------------------
+* Add ideal imu sensor for the Crazyflie
+* Contributors: Ria Sonecha, Giuseppe Silano
+
 4.0.9 (2020-03-22)
 ------------------